+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Added progress tracker component and documentation.
 - Added tooltip component.
 - Added media object component.
+- Added action list component.
 
 ### Changed